requirements.txt
pandas==0.20.3
PyMySQL==0.7.11
h5py==2.7.1
keras==2.0.8
sqlalchemy==1.1.14
requests==2.18.4
tensorflow==1.3.0
PyYAML==3.12
tensorboard==1.0.0a6
